requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
nltk==3.9.1
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
regex==2024.11.6
six==1.16.0
sqlparse==0.5.0
textblob==0.18.0.post0
textblob-fr==0.2.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
